requirements.txt
httplib2==0.9
oauth2==1.5.211
py-trello==0.1.4
pypodio2==0.1
requests==2.3.0
wsgiref==0.1.2
